Data-Science/Agents-google/pyproject.toml
[project]
name = "agents-google"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "google-adk>=1.27.5",
    "httpx>=0.28.1",
    "pexpect>=4.9.0",
]
